Updated to R-devel revision 76888.

+++ R/src/library/tools/R/install.R
@@ -130,6 +130,12 @@ if(FALSE) {
     quote_path <- function(path, quote = "'")
     	paste0(quote, gsub("\\", "\\\\", path, fixed=TRUE), quote)
 
+    # Escape backslashes in a replacement string for gsub etc.
+    # To be used when the replacement is a path name which may include
+    # backslashes, e.g. with UNC paths on Windows.
+    quote_replacement <- function(r)
+        paste0(gsub("\\", "\\\\", r, fixed=TRUE))
+
     on.exit(do_exit_on_error())
     WINDOWS <- .Platform$OS.type == "windows"
 
@@ -628,7 +634,8 @@ if(FALSE) {
                     idxs <- gsub(re, "\\1", out)
                     old_paths <- paths
                     # "\\$ORIGIN/.."
-                    paths <- gsub(instdir, final_instdir, paths, fixed = TRUE)
+                    paths <- gsub(instdir, quote_replacement(final_instdir),
+                                  paths, fixed = TRUE)
                     changed <- paths != old_paths
                     paths <- paths[changed]
                     old_paths <- old_paths[changed]
@@ -663,8 +670,8 @@ if(FALSE) {
                         grepl(instdir, oldid, fixed = TRUE)) {
 
                         hardcoded_paths <- TRUE
-                        newid <- gsub(instdir, final_instdir, oldid,
-                                      fixed = TRUE)
+                        newid <- gsub(instdir, quote_replacement(final_instdir),
+                                      oldid, fixed = TRUE)
                         cmd <- paste("install_name_tool -id", shQuote(newid),
                                      shQuote(l))
                         message(cmd)
@@ -684,8 +691,8 @@ if(FALSE) {
                                   paths)
                     old_paths <- paths
                     # "@loader_path/.."
-                    paths <- gsub(instdir, final_instdir, paths,
-                                  fixed = TRUE)
+                    paths <- gsub(instdir, quote_replacement(final_instdir),
+                                  paths, fixed = TRUE)
                     changed <- paths != old_paths
                     paths <- paths[changed]
                     old_paths <- old_paths[changed]
@@ -719,8 +726,8 @@ if(FALSE) {
                         paths <- gsub("(.*) \\(offset .*", "\\1", paths)
                         old_paths <- paths
                         # "@loader_path/.."
-                        paths <- gsub(instdir, final_instdir, paths,
-                                               fixed = TRUE)
+                        paths <- gsub(instdir, quote_replacement(final_instdir),
+                                      paths, fixed = TRUE)
                         changed <- paths != old_paths
                         paths <- paths[changed]
                         old_paths <- old_paths[changed]
@@ -754,8 +761,8 @@ if(FALSE) {
                                intern = TRUE))
                     old_rpath <- rpath
                     # "\\$ORIGIN/.."
-                    rpath <- gsub(instdir, final_instdir, rpath,
-                                  fixed = TRUE)
+                    rpath <- gsub(instdir, quote_replacement(final_instdir),
+                                  rpath, fixed = TRUE)
                     if (length(rpath) && nzchar(rpath) && old_rpath != rpath) {
                         hardcoded_paths <- TRUE
                         cmd <- paste("patchelf", "--set-rpath",
@@ -780,8 +787,8 @@ if(FALSE) {
                         paths <- gsub(re, "\\1", out)
                         old_paths <- paths
                         # "\\$ORIGIN/.."
-                        paths <- gsub(instdir, final_instdir, paths,
-                                      fixed = TRUE)
+                        paths <- gsub(instdir, quote_replacement(final_instdir),
+                                      paths, fixed = TRUE)
                         changed <- paths != old_paths
                         paths <- paths[changed]
                         old_paths <- old_paths[changed]
@@ -816,7 +823,8 @@ if(FALSE) {
                     rpath <- gsub(".*PATH=", "", rpath)
                     old_rpath <- rpath
                     # "\\$ORIGIN/.."
-                    rpath <- gsub(instdir, final_instdir, rpath, fixed = TRUE)
+                    rpath <- gsub(instdir, quote_replacement(final_instdir),
+                                  rpath, fixed = TRUE)
                     if (length(rpath) && nzchar(rpath) && old_rpath != rpath) {
                         hardcoded_paths <- TRUE
                         cmd <- paste("chrpath", "-r", shQuote(rpath),
@@ -854,7 +862,8 @@ if(FALSE) {
                     }
                 }
             }
-        }
+        } # patch_rpaths()
+
         ## Make the destination directories available to the developer's
         ## installation scripts (e.g. configure)
         Sys.setenv(R_LIBRARY_DIR = lib)
@@ -1066,7 +1075,7 @@ if(FALSE) {
                                           target, sQuote(pkgname)),
                                  call. = FALSE, domain = NA)
                     }
-                    clink_cppflags <- paste(paste0('-I"', paths, '/include"'),
+                    clink_cppflags <- paste(paste0("-I'", paths, "/include'"),
                                             collapse = " ")
                     Sys.setenv(CLINK_CPPFLAGS = clink_cppflags)
                 }
@@ -1401,7 +1410,7 @@ if(FALSE) {
             } else character()
             for(e in ignore)
                 i_dirs <- filtergrep(e, i_dirs, perl = TRUE, ignore.case = TRUE)
-            lapply(gsub("^inst", instdir, i_dirs),
+            lapply(gsub("^inst", quote_replacement(instdir), i_dirs),
                    function(p) dir.create(p, FALSE, TRUE)) # be paranoid
             i_files <- list.files("inst", all.files = TRUE,
                                   full.names = TRUE, recursive = TRUE)
@@ -1417,7 +1426,7 @@ if(FALSE) {
                 i_files <- filtergrep("inst/doc/.*[.](png|jpg|jpeg|gif|ps|eps)$",
                                       i_files, perl = TRUE, ignore.case = TRUE)
             i_files <- i_files %w/o% "Makefile"
-            i2_files <- gsub("^inst", instdir, i_files)
+            i2_files <- gsub("^inst", quote_replacement(instdir), i_files)
             file.copy(i_files, i2_files)
             if (!WINDOWS) {
                 ## make executable if the source file was (for owner)
@@ -1495,7 +1504,7 @@ if(FALSE) {
                 set.install.dir <- ""
             cmd <- append(cmd,
                 paste0("tools:::makeLazyLoading(\"", pkg_name, "\", ",
-                                                    "\"", lib, "\", ",
+                                              quote_path(lib), ", ",
                                 "keep.source = ", keep.source, ", ",
                         "keep.parse.data = ", keep.parse.data,
                                               set.install.dir, ")"))
@@ -1631,10 +1640,15 @@ if(FALSE) {
             if (WINDOWS) {
                 unlink(final_instdir, recursive = TRUE) # needed for file.rename
                 if (!file.rename(instdir, final_instdir)) {
-                    message("WARNING: moving package to final location failed, copying instead")
-                    file.copy(instdir, dirname(final_instdir), recursive = TRUE,
-                              copy.date = TRUE)
-                    unlink(instdir, recursive = TRUE)
+                    if (dir.exists(instdir) && !dir.exists(final_instdir)) {
+                        message("WARNING: moving package to final location failed, copying instead")
+                        ret <- file.copy(instdir, dirname(final_instdir),
+                                         recursive = TRUE, copy.date = TRUE)
+                        if (any(!ret))
+                            errmsg("   copying to final location failed")
+                        unlink(instdir, recursive = TRUE)
+                    } else
+                        errmsg("   moving to final location failed")
                 }
             } else {
                 patch_rpaths()
@@ -1697,7 +1711,7 @@ if(FALSE) {
             starsmsg(stars, ngettext(length(so_s), "stripping dynamic library under lib", "stripping dynamic libraries under lib", domain = "R-tools"))
             system(paste(c(strip_cmd, shQuote(so_s)), collapse = " "))
         }
-    }
+    } ## do_install_source
 
     options(showErrorCalls = FALSE)
     pkgs <- character()
